# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://localhost:4005/"
    },
    {
      "url": "https://localhost:4005/"
    }
  ],
  "paths": {
    "/api/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/health-check": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/users/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {},
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/verify-reset-token": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/{id}/reset-password": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "ref": "#/components/schemas/CreateUser"
              }
            },
            "application/xml": {
              "schema": {
                "ref": "#/components/schemas/CreateUser"
              }
            }
          }
        }
      }
    },
    "/api/users/forgot-password": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPassword"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPassword"
              }
            }
          }
        }
      }
    },
    "/api/auth/user": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePassword"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/ChangePassword"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "ChangePassword": {
        "type": "object",
        "properties": {
          "old_password": {
            "type": "string",
            "title": "Old Password"
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[\\d])(?=.*[!@#$%^&*])(?=.*[A-Z])[a-zA-Z0-9!@#$%^&*]{8,}$",
            "title": "Password"
          },
          "confirm_password": {
            "type": "string",
            "title": "Confirm Password"
          }
        },
        "required": [
          "old_password",
          "password",
          "confirm_password"
        ],
        "additionalProperties": false
      },
      "CreateUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 225,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[\\d])(?=.*[!@#$%^&*])(?=.*[A-Z])[a-zA-Z0-9!@#$%^&*]{8,}$",
            "title": "Password"
          },
          "confirm_password": {
            "type": "string",
            "title": "Confirm Password"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "title": "Phone"
          },
          "status": {
            "type": "boolean",
            "title": "Status"
          },
          "image_url": {
            "type": "string",
            "title": "Image"
          },
          "reset_token": {
            "type": "string",
            "title": "Token"
          },
          "role_id": {
            "type": "number",
            "format": "float",
            "title": "Role"
          },
          "created_by": {
            "type": "number",
            "format": "float",
            "title": "Created By"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "confirm_password",
          "created_by"
        ],
        "additionalProperties": false
      },
      "UpdateUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 225,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "title": "Phone"
          },
          "status": {
            "type": "boolean",
            "title": "Status"
          },
          "image_url": {
            "type": "string",
            "title": "Image"
          },
          "role_id": {
            "type": "number",
            "format": "float",
            "title": "Role"
          },
          "updated_by": {
            "type": "number",
            "format": "float",
            "title": "Updated By"
          }
        },
        "required": [
          "updated_by"
        ],
        "additionalProperties": false
      },
      "ForgotPassword": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "base_url": {
            "type": "string",
            "title": "URL"
          }
        },
        "required": [
          "email",
          "base_url"
        ],
        "additionalProperties": false
      },
      "ResetPassword": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "title": "Reset Token"
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[\\d])(?=.*[!@#$%^&*])(?=.*[A-Z])[a-zA-Z0-9!@#$%^&*]{8,}$",
            "title": "Password"
          },
          "confirm_password": {
            "type": "string",
            "title": "Confirm Password"
          }
        },
        "required": [
          "token",
          "password",
          "confirm_password"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}